Dev martinez

+++ services/comentario.service.js
@@ -8,11 +8,22 @@ const getAllComentarios = async () => {
 }
 
 const createComentario = async (newComentario) => {
+  const { comentario } = newComentario;
+  const existeComentario = await Comentario.findOne({
+    where: { comentario }
+  });
+  if (existeComentario) {
+    throw new Error('El comentario ya existe en la base de datos');
+  }
   const comentarioCreated = await Comentario.create(newComentario);
   return comentarioCreated;
 }
 
 const updateComentario = async (comentario, id) => {
+  const comentarioExiste = await Comentario.findByPk(id);
+  if (!comentarioExiste) {
+    throw new Error('El comentario con el ID proporcionado no existe en la base de datos');
+  }
   const comentarioUpdated = await Comentario.update(comentario, {
     where: {
       comentario_id: id
@@ -22,13 +33,20 @@ const updateComentario = async (comentario, id) => {
 }
 
 const deleteComentario = async (id) => {
-  const comentarioDeleted = await Comentario.destroy({
+  const comentario = await validacionExisteComentario(id);
+  const valoraciones = await comentario.getValoraciones();
+  if (valoraciones.length > 0) {
+    throw new Error('No se puede eliminar el comentario porque tiene valoraciones asociadas');
+  }
+
+  const comentariosEliminados = await Comentario.destroy({
     where: {
       comentario_id: id
     }
   });
-  return comentarioDeleted;
+  return comentariosEliminados;
 }
+
 
 module.exports = {
   getAllComentarios,
